one-nav - move dock-option and side-option out
